refactored all redux slices to common structure
