--- cx-portal/src/state/deprecated/apps/appsAPI.ts
@@ -1,5 +1,0 @@
-import { createAction } from '@reduxjs/toolkit'
-
-export const appsCallBegan = createAction('apps/callBegan')
-export const appsCallSuccess = createAction('apps/callSuccess')
-export const appsCallFailed = createAction('apps/callFailed')
