--- cx-portal/src/state/deprecated/appstore/AppstoreTypes.ts
@@ -1,18 +1,0 @@
-export interface AppstoreImage {
-  src: string
-  alt?: string
-}
-
-export type MouseEventHandler = (e: MouseEvent) => void
-
-export interface AppstoreApp {
-  id: string
-  title: string
-  subtitle: string
-  description: string
-  price: string
-  rating?: number
-  image?: AppstoreImage
-  onButtonClick?: MouseEventHandler
-  onSecondaryButtonClick?: MouseEventHandler
-}
