--- cx-portal/src/state/features/middleware.ts
@@ -1,8 +1,0 @@
-import appsMW from '../deprecated/apps/appsMW'
-
-const rootMiddleware = (getDefaultMiddleware: any) => [
-  ...getDefaultMiddleware(),
-  appsMW,
-]
-
-export default rootMiddleware
