+++ cx-portal/src/state/features/partnerNetwork/actions.ts
@@ -1,15 +1,13 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import { api } from 'state/api'
-import { SearchParams } from 'types/partnerNetwork/PartnerNetworkTypes'
+import { SearchParams } from './types'
+import { PartnerNetworkApi } from './api'
 
 const fetchBusinessPartners = createAsyncThunk(
   'partnerNetwork/fetchBusinessPartners',
   async ({ params, token }: { params: SearchParams; token: string }) => {
     try {
       // Call axios instance to get values
-
-      const partnerNetworkApi = api.PartnerNetworkApi.getInstance(token)
-      return await partnerNetworkApi.getAllBusinessPartner(params)
+      return await PartnerNetworkApi.getInstance().getAllBusinessPartner(params)
     } catch (error: unknown) {
       console.error('api call error:', error)
       throw Error('fetchBusinessPartners api call error')
@@ -22,8 +20,7 @@ const getOneBusinessPartner = createAsyncThunk(
   async ({ bpn, token }: { bpn: string; token: string }) => {
     try {
       // Call axios instance to get values
-      const partnerNetworkApi = api.PartnerNetworkApi.getInstance(token)
-      return await partnerNetworkApi.getBusinessPartnerByBpn(bpn)
+      return await PartnerNetworkApi.getInstance().getBusinessPartnerByBpn(bpn)
     } catch (error: unknown) {
       console.error('api call error:', error)
       throw Error('getOneBusinessPartner api call error')
