+++ cx-portal/src/state/features/reducer.ts
@@ -1,7 +1,6 @@
 import { combineReducers } from 'redux'
-import appsSlice from 'state/deprecated/apps/apps'
-import userSlice from 'state/features/user/userSlice'
-import partnerNetworkSlice from 'state/features/partnerNetwork/partnerNetworkSlice'
+import userSlice from 'state/features/user/slice'
+import partnerNetworkSlice from 'state/features/partnerNetwork/slice'
 import userAdministrationSlice from 'state/features/userAdministration/slice'
 import appMarketplaceSlice from 'state/features/appMarketplace/slice'
 import newsSlice from './news/slice'
@@ -9,7 +8,6 @@ import licensesSlice from './licenses/slice'
 
 // Reducers need separate export for testing library
 export const reducers = {
-  apps: appsSlice.reducer,
   user: userSlice,
   news: newsSlice.reducer,
   licenses: licensesSlice.reducer,
