+++ cx-portal/src/state/features/user/slice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { RootState } from 'state/store'
-import { IUser } from 'types/user/UserTypes'
+import { IUser } from './types'
 
 const initialState: IUser = {
   userName: '',
@@ -25,6 +25,6 @@ const userSlice = createSlice({
 })
 
 export const { setLoggedUser } = userSlice.actions
-export const selectorUser = (state: RootState): IUser => state.user
+export const userSelector = (state: RootState): IUser => state.user
 
 export default userSlice.reducer
