+++ cx-portal/src/state/store.ts
@@ -1,10 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit'
-import rootMiddleware from './features/middleware'
 import rootReducer from './features/reducer'
 
 export const store = configureStore({
   reducer: rootReducer,
-  middleware: rootMiddleware,
 })
 
 type RootState = ReturnType<typeof store.getState>
