+++ cx-portal/src/utils/dataMapper.test.ts
@@ -9,7 +9,7 @@ import TestSingleDataGridData from '../../public/testdata/partnerNetwork/busines
 import {
   BusinessPartner,
   BusinessPartnerResponse,
-} from 'types/partnerNetwork/PartnerNetworkTypes'
+} from 'state/features/partnerNetwork/types'
 
 describe('DataMapperTest', () => {
   it('mapBusinessPartnerToDataGrid maps correctly', () => {
